Initial port to 1.14.4.  Runs!

+++ src/main/java/com/wumple/megamap/api/IMegaMapAPI.java
@@ -3,12 +3,11 @@ package com.wumple.megamap.api;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
-import net.minecraft.world.storage.MapData;
 
 public interface IMegaMapAPI
 {
 
-    public ItemStack setupNewMap(World worldIn, double worldX, double worldZ, byte scale, boolean trackingPosition, boolean unlimitedTracking);
+    public ItemStack setupNewMap(World worldIn, int worldX, int worldZ, byte scale, boolean trackingPosition, boolean unlimitedTracking);
 
     boolean isEmptyMap(ItemStack itemstack1);
 
@@ -18,9 +17,11 @@ public interface IMegaMapAPI
 
     ItemStack copyMap(ItemStack itemstack, int i);
 
+    /*
     boolean isExplorationMap(MapData mapData);
 
     boolean isExplorationMap(ItemStack itemstack, World worldIn);
+    */
 
     Item getFilledMegaMapItem();
